+++ src/ui/marker.ts
@@ -126,6 +126,7 @@ export class Marker extends Evented {
     _popup: Popup;
     _lngLat: LngLat;
     _pos: Point;
+    _flatPos: Point;
     _color: string;
     _scale: number;
     _defaultMarker: boolean;
@@ -568,10 +569,14 @@ export class Marker extends Evented {
         }
 
         if (this._map.transform.renderWorldCopies) {
-            this._lngLat = smartWrap(this._lngLat, this._pos, this._map.transform);
+            this._lngLat = smartWrap(this._lngLat, this._flatPos, this._map.transform);
         }
 
-        this._pos = this._map.project(this._lngLat)._add(this._offset);
+        this._flatPos = this._pos = this._map.project(this._lngLat)._add(this._offset);
+        if (this._map.terrain) {
+            // flat position is saved because smartWrap needs non-elevated points
+            this._flatPos = this._map.transform.locationPoint(this._lngLat)._add(this._offset);
+        }
 
         let rotation = '';
         if (this._rotationAlignment === 'viewport' || this._rotationAlignment === 'auto') {