+++ src/ui/popup.ts
@@ -158,6 +158,7 @@ export class Popup extends Evented {
     _lngLat: LngLat;
     _trackPointer: boolean;
     _pos: Point;
+    _flatPos: Point;
 
     constructor(options?: PopupOptions) {
         super();
@@ -570,12 +571,16 @@ export class Popup extends Evented {
         }
 
         if (this._map.transform.renderWorldCopies && !this._trackPointer) {
-            this._lngLat = smartWrap(this._lngLat, this._pos, this._map.transform);
+            this._lngLat = smartWrap(this._lngLat, this._flatPos, this._map.transform);
         }
 
         if (this._trackPointer && !cursor) return;
 
-        const pos = this._pos = this._trackPointer && cursor ? cursor : this._map.project(this._lngLat);
+        const pos = this._flatPos = this._pos = this._trackPointer && cursor ? cursor : this._map.project(this._lngLat);
+        if (this._map.terrain) {
+            // flat position is saved because smartWrap needs non-elevated points
+            this._flatPos = this._trackPointer && cursor ? cursor : this._map.transform.locationPoint(this._lngLat);
+        }
 
         let anchor = this.options.anchor;
         const offset = normalizeOffset(this.options.offset);