+++ CHANGELOG.md
@@ -5,7 +5,12 @@
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
+
 - Fix Style.setState ignoring validate flag.
+
+- Fix marker flying off near horizon ([3704](https://github.com/maplibre/maplibre-gl-js/pull/3704))
+- _...Add new stuff here..._
+
 
 ## 4.0.1