+++ CHANGELOG.md
@@ -1,14 +1,20 @@
 ## main
 
 ### ✨ Features and improvements
-
 - _...Add new stuff here..._
+
+### 🐞 Bug fixes
+- _...Add new stuff here..._
+
+## 4.0.2
+
+### ✨ Features and improvements
+
 
 ### 🐞 Bug fixes
 
 - Fix `Style.setState` ignoring validate flag ([#3709](https://github.com/maplibre/maplibre-gl-js/pull/3709))
 - Fix marker flying off near horizon ([3704](https://github.com/maplibre/maplibre-gl-js/pull/3704))
-- _...Add new stuff here..._
 
 ## 4.0.1